replace `{int,uint}` with `{isize,usize}`

+++ src/lru_cache.rs
@@ -19,7 +19,7 @@
 //! ```rust
 //! use collect::LruCache;
 //!
-//! let mut cache: LruCache<int, int> = LruCache::new(2);
+//! let mut cache: LruCache<isize, isize> = LruCache::new(2);
 //! cache.insert(1, 10);
 //! cache.insert(2, 20);
 //! cache.insert(3, 30);
@@ -50,7 +50,7 @@ use linked_hash_map::LinkedHashMap;
 /// An LRU Cache.
 pub struct LruCache<K, V> {
     map: LinkedHashMap<K, V>,
-    max_size: uint,
+    max_size: usize,
 }
 
 impl<K: Hash<HmHasher> + Eq, V> LruCache<K, V> {
@@ -60,10 +60,10 @@ impl<K: Hash<HmHasher> + Eq, V> LruCache<K, V> {
     ///
     /// ```rust
     /// use collect::LruCache;
-    /// let mut cache: LruCache<int, &str> = LruCache::new(10);
+    /// let mut cache: LruCache<isize, &str> = LruCache::new(10);
     /// ```
     #[unstable = "matches collection reform specification, waiting for dust to settle"]
-    pub fn new(capacity: uint) -> LruCache<K, V> {
+    pub fn new(capacity: usize) -> LruCache<K, V> {
         LruCache {
             map: LinkedHashMap::new(),
             max_size: capacity,
@@ -152,17 +152,17 @@ impl<K: Hash<HmHasher> + Eq, V> LruCache<K, V> {
     ///
     /// ```rust
     /// use collect::LruCache;
-    /// let mut cache: LruCache<int, &str> = LruCache::new(2);
+    /// let mut cache: LruCache<isize, &str> = LruCache::new(2);
     /// assert_eq!(cache.capacity(), 2);
     /// ```
     #[unstable = "matches collection reform specification, waiting for dust to settle"]
-    pub fn capacity(&self) -> uint {
+    pub fn capacity(&self) -> usize {
         self.max_size
     }
 
     /// Deprecated: Renamed to `set_capacity`.
     #[deprecated = "Renamed to `set_capacity`"]
-    pub fn change_capacity(&mut self, capacity: uint) {
+    pub fn change_capacity(&mut self, capacity: usize) {
         self.set_capacity(capacity)
     }
 
@@ -198,7 +198,7 @@ impl<K: Hash<HmHasher> + Eq, V> LruCache<K, V> {
     /// assert_eq!(cache.get(&3), Some(&"c"));
     /// ```
     #[unstable = "matches collection reform specification, waiting for dust to settle"]
-    pub fn set_capacity(&mut self, capacity: uint) {
+    pub fn set_capacity(&mut self, capacity: usize) {
         for _ in range(capacity, self.len()) {
             self.remove_lru();
         }
@@ -212,7 +212,7 @@ impl<K: Hash<HmHasher> + Eq, V> LruCache<K, V> {
 
     /// Return the number of key-value pairs in the cache.
     #[unstable = "matches collection reform specification, waiting for dust to settle"]
-    pub fn len(&self) -> uint { self.map.len() }
+    pub fn len(&self) -> usize { self.map.len() }
 
     /// Returns whether the cache is currently empty.
     #[unstable = "matches collection reform specification, waiting for dust to settle"]
@@ -262,7 +262,7 @@ mod tests {
 
     #[test]
     fn test_put_and_get() {
-        let mut cache: LruCache<int, int> = LruCache::new(2);
+        let mut cache: LruCache<isize, isize> = LruCache::new(2);
         cache.insert(1, 10);
         cache.insert(2, 20);
         assert_opt_eq(cache.get(&1), 10);
@@ -293,7 +293,7 @@ mod tests {
 
     #[test]
     fn test_pop() {
-        let mut cache: LruCache<int, int> = LruCache::new(2);
+        let mut cache: LruCache<isize, isize> = LruCache::new(2);
         cache.insert(1, 10);
         cache.insert(2, 20);
         assert_eq!(cache.len(), 2);
@@ -306,7 +306,7 @@ mod tests {
 
     #[test]
     fn test_change_capacity() {
-        let mut cache: LruCache<int, int> = LruCache::new(2);
+        let mut cache: LruCache<isize, isize> = LruCache::new(2);
         assert_eq!(cache.capacity(), 2);
         cache.insert(1, 10);
         cache.insert(2, 20);
@@ -334,7 +334,7 @@ mod tests {
 
     #[test]
     fn test_remove() {
-        let mut cache: LruCache<int, int> = LruCache::new(3);
+        let mut cache: LruCache<isize, isize> = LruCache::new(3);
         cache.insert(1, 10);
         cache.insert(2, 20);
         cache.insert(3, 30);
@@ -355,7 +355,7 @@ mod tests {
 
     #[test]
     fn test_clear() {
-        let mut cache: LruCache<int, int> = LruCache::new(2);
+        let mut cache: LruCache<isize, isize> = LruCache::new(2);
         cache.insert(1, 10);
         cache.insert(2, 20);
         cache.clear();
